move transformTaslFromString and transformImage to common
