+++ common/services/prismic/transformers/index.ts
@@ -1,0 +1,41 @@
+import { Tasl } from '../../../model/tasl';
+import { licenseTypeArray } from '../../../model/license';
+
+export function transformTaslFromString(pipedString: string | null): Tasl {
+  if (pipedString === null) {
+    return { title: '' };
+  }
+
+  // We expect a string of title|author|sourceName|sourceLink|license|copyrightHolder|copyrightLink
+  // e.g. Self|Rob Bidder|||CC-BY-NC
+  try {
+    const list = (pipedString || '').split('|');
+    const v = list
+      .concat(Array(7 - list.length))
+      .map(v => (!v.trim() ? undefined : v.trim()));
+
+    const [
+      title,
+      author,
+      sourceName,
+      sourceLink,
+      maybeLicense,
+      copyrightHolder,
+      copyrightLink,
+    ] = v;
+    const license = licenseTypeArray.find(l => l === maybeLicense);
+    return {
+      title,
+      author,
+      sourceName,
+      sourceLink,
+      license,
+      copyrightHolder,
+      copyrightLink,
+    };
+  } catch (e) {
+    return {
+      title: pipedString,
+    };
+  }
+}
