+++ content/webapp/services/prismic/transformers/images.ts
@@ -1,15 +1,12 @@
 import { Image, PromoSliceZone } from '../types';
-import {
-  EmptyImageFieldImage,
-  FilledImageFieldImage,
-  RichTextField,
-} from '@prismicio/types';
+import { RichTextField } from '@prismicio/types';
 import { CaptionedImage } from '@weco/common/model/captioned-image';
 import isEmptyObj from '@weco/common/utils/is-empty-object';
 import { ImageType } from '@weco/common/model/image';
 import { ImagePromo } from '../../../types/image-promo';
-import { asRichText, asText, transformTaslFromString } from '.';
+import { asRichText, asText } from '.';
 import * as prismicT from '@prismicio/types';
+import { transformImage } from '@weco/common/services/prismic/transformers/images';
 
 export const placeHolderImage: ImageType = {
   contentUrl: 'https://via.placeholder.com/1600x900?text=%20',
@@ -58,50 +55,6 @@ export function transformPromoToCaptionedImage(
   } else {
     return undefined;
   }
-}
-
-// when images have crops, event if the image isn't attached, we get e.g.
-// { '32:15': {}, '16:9': {}, square: {} }
-function isImageLink(
-  maybeImage: EmptyImageFieldImage | FilledImageFieldImage
-): maybeImage is FilledImageFieldImage {
-  return Boolean(maybeImage && maybeImage.dimensions);
-}
-
-export function transformImage(
-  maybeImage: EmptyImageFieldImage | FilledImageFieldImage | undefined
-): ImageType | undefined {
-  return maybeImage && isImageLink(maybeImage)
-    ? transformFilledImage(maybeImage)
-    : undefined;
-}
-
-// These are the props returned on a prismic image object
-const prismicImageProps = ['dimensions', 'alt', 'copyright', 'url'];
-
-// We don't export this, as we probably always want to check ^ first
-function transformFilledImage(image: FilledImageFieldImage): ImageType {
-  const tasl = transformTaslFromString(image.copyright);
-  const crops = Object.keys(image)
-    .filter(key => prismicImageProps.indexOf(key) === -1)
-    .filter(key => isImageLink(image[key]))
-    .map(key => ({
-      key,
-      image: transformFilledImage(image[key]),
-    }))
-    .reduce((acc, { key, image }) => {
-      acc[key] = image;
-      return acc;
-    }, {});
-
-  return {
-    contentUrl: image.url,
-    width: image.dimensions.width,
-    height: image.dimensions.height,
-    alt: image.alt!,
-    tasl: tasl,
-    crops: crops,
-  };
 }
 
 // null is valid to use the default image,
