+++ content/webapp/services/prismic/transformers/index.ts
@@ -54,9 +54,8 @@ import {
   transformTextSlice,
   transformTitledTextListSlice,
 } from './body';
-import { transformImage, transformImagePromo } from './images';
-import { Tasl } from '@weco/common/model/tasl';
-import { licenseTypeArray } from '@weco/common/model/license';
+import { transformImage } from '@weco/common/services/prismic/transformers/images';
+import { transformImagePromo } from './images';
 import { WithPageFormat } from '../types/pages';
 import { WithEventFormat } from '../types/events';
 import { Format } from '../../../types/format';
@@ -425,43 +424,4 @@ export function transformGenericFields(doc: Doc): GenericContentFields {
     // TODO: find a way to enforce this.
     labels: [],
   };
-}
-
-export function transformTaslFromString(pipedString: string | null): Tasl {
-  if (pipedString === null) {
-    return { title: '' };
-  }
-
-  // We expect a string of title|author|sourceName|sourceLink|license|copyrightHolder|copyrightLink
-  // e.g. Self|Rob Bidder|||CC-BY-NC
-  try {
-    const list = (pipedString || '').split('|');
-    const v = list
-      .concat(Array(7 - list.length))
-      .map(v => (!v.trim() ? undefined : v.trim()));
-
-    const [
-      title,
-      author,
-      sourceName,
-      sourceLink,
-      maybeLicense,
-      copyrightHolder,
-      copyrightLink,
-    ] = v;
-    const license = licenseTypeArray.find(l => l === maybeLicense);
-    return {
-      title,
-      author,
-      sourceName,
-      sourceLink,
-      license,
-      copyrightHolder,
-      copyrightLink,
-    };
-  } catch (e) {
-    return {
-      title: pipedString,
-    };
-  }
 }
